Add user-related handlers and routes for fetching user statistics and feedback

+++ app/services/NavigationService.ts
@@ -28,14 +28,15 @@ export class NavigationService {
             throw error;
         }
     }
-    static async getSuccessfulNavigations(){
-        try{
-            const result = await prisma.$queryRaw<{date: Date; total_attempts: number; successful_navigations: number; success_rate: number}[]>
-            `
+    static async getSuccessfulNavigations() {
+        try {
+            const result = await prisma.$queryRaw<
+                { date: Date; total_attempts: number; successful_navigations: number; success_rate: number }[]
+            >`
                 SELECT
                 DATE_TRUNC('day', start_time) AS date,
-                COUNT(*) AS total_attempts,
-                SUM(CASE WHEN end_time IS NOT NULL AND rerouting_count = 0 THEN 1 ELSE 0 END) AS successful_navigations,
+                COUNT(*)::INTEGER AS total_attempts,
+                SUM(CASE WHEN end_time IS NOT NULL AND rerouting_count = 0 THEN 1 ELSE 0 END)::INTEGER AS successful_navigations,
                 ROUND((SUM(CASE WHEN end_time IS NOT NULL AND rerouting_count = 0 THEN 1 ELSE 0 END)::NUMERIC / NULLIF(COUNT(*), 0)) * 100, 2) AS success_rate
                 FROM 
                 "Navigation_logs"
@@ -44,11 +45,11 @@ export class NavigationService {
                 ORDER BY date;
             `;
             return result;
-        }
-        catch(error){
+        } catch (error) {
             console.error("Error fetching successful navigations:", error);
             throw error;
         }
     }
+    
 
 }
